+++ persistent/Database/Persist/Sql/Migration.hs
@@ -5,8 +5,10 @@ module Database.Persist.Sql.Migration
   , showMigration
   , getMigration
   , runMigration
+  , runMigrationQuiet
   , runMigrationSilent
   , runMigrationUnsafe
+  , runMigrationUnsafeQuiet
   , migrate
   -- * Utilities for constructing migrations
   , reportErrors
@@ -16,6 +18,7 @@ module Database.Persist.Sql.Migration
   ) where
 
 
+import Control.Exception (throwIO)
 import Control.Monad (liftM, unless)
 import Control.Monad.IO.Unlift
 import Control.Monad.Trans.Class (MonadTrans (..))
@@ -80,8 +83,26 @@ runMigration :: MonadIO m
              -> ReaderT SqlBackend m ()
 runMigration m = runMigration' m False >> return ()
 
+-- | Same as 'runMigration', but does not report the individual migrations on
+-- stderr. Instead it returns a list of the executed SQL commands.
+--
+-- This is a safer/more robust alternative to 'runMigrationSilent', but may be
+-- less silent for some persistent implementations, most notably
+-- persistent-postgresql
+--
+-- @since 2.10.2
+runMigrationQuiet :: MonadIO m
+                  => Migration
+                  -> ReaderT SqlBackend m [Text]
+runMigrationQuiet m = runMigration' m True
+
 -- | Same as 'runMigration', but returns a list of the SQL commands executed
 -- instead of printing them to stderr.
+--
+-- This function silences the migration by remapping 'stderr'. As a result, it
+-- is not thread-safe and can clobber output from other parts of the program.
+-- This implementation method was chosen to also silence postgresql migration
+-- output on stderr, but is not recommended!
 runMigrationSilent :: MonadUnliftIO m
                    => Migration
                    -> ReaderT SqlBackend m [Text]
@@ -99,7 +120,7 @@ runMigration'
 runMigration' m silent = do
     mig <- parseMigration' m
     if any fst mig
-        then error $ concat
+        then liftIO . throwIO . PersistError . pack $ concat
                  [ "\n\nDatabase migration: manual intervention required.\n"
                  , "The unsafe actions are prefixed by '***' below:\n\n"
                  , unlines $ map displayMigration mig
@@ -115,9 +136,24 @@ runMigration' m silent = do
 runMigrationUnsafe :: MonadIO m
                    => Migration
                    -> ReaderT SqlBackend m ()
-runMigrationUnsafe m = do
+runMigrationUnsafe m = runMigrationUnsafe' False m >> return ()
+
+-- | Same as 'runMigrationUnsafe', but returns a list of the SQL commands
+-- executed instead of printing them to stderr.
+--
+-- @since 2.10.2
+runMigrationUnsafeQuiet :: MonadIO m
+                        => Migration
+                        -> ReaderT SqlBackend m [Text]
+runMigrationUnsafeQuiet = runMigrationUnsafe' True
+
+runMigrationUnsafe' :: MonadIO m
+                    => Bool
+                    -> Migration
+                    -> ReaderT SqlBackend m [Text]
+runMigrationUnsafe' silent m = do
     mig <- parseMigration' m
-    mapM_ (executeMigrate False) $ sortMigrations $ allSql mig
+    mapM (executeMigrate silent) $ sortMigrations $ allSql mig
 
 executeMigrate :: MonadIO m => Bool -> Text -> ReaderT SqlBackend m Text
 executeMigrate silent s = do